# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-market investment portfolio tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portfolios/{portfolioId}/csv/import:
    post:
      operationId: importCsv
      tags: [csv]
      summary: Import transactions from CSV text
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CsvImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvImportResult"
  /portfolios/{portfolioId}/csv/export:
    get:
      operationId: exportCsv
      tags: [csv]
      summary: Export transactions as CSV text
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CSV content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsvExportResult"

  # ── WATCHLISTS ────────────────────────────────────────────────────────────────
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List all price alerts
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceAlert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a price alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceAlertInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlert"
  /alerts/{alertId}:
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update or dismiss an alert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceAlertUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlert"
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Delete a price alert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────────────
...
  /prices/search:
    get:
      operationId: searchSymbols
      tags: [prices]
      summary: Search for symbols by query string
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: market
          in: query
          schema:
            type: string
            enum: [PSE, LSE, US, CRYPTO, CUSTOM]
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SymbolSearchResult"

  # ── DIVIDEND CALENDAR ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CsvImportInput:
      type: object
      required: [csvContent]
      properties:
        csvContent:
          type: string
        dryRun:
          type: boolean
          default: false
    CsvImportResult:
      type: object
      required: [imported, skipped, errors]
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
    CsvExportResult:
      type: object
      required: [csvContent, filename]
      properties:
        csvContent:
          type: string
        filename:
          type: string

    # ── Watchlist ──────────────────────────────────────────────────────────────
...
    PriceAlert:
      type: object
      required: [id, userId, symbol, market, name, condition, targetPrice, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        symbol:
          type: string
        market:
          type: string
          enum: [PSE, LSE, US, CRYPTO, CUSTOM]
        name:
          type: string
        condition:
          type: string
          enum: [GTE, LTE, EQ, CROSS_UP, CROSS_DOWN]
        targetPrice:
          type: number
        currentPrice:
          type: ["number", "null"]
        status:
          type: string
          enum: [PENDING, FIRED, CLEARED, DISMISSED]
        currency:
          type: ["string", "null"]
        firedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    PriceAlertInput:
      type: object
      required: [symbol, market, name, condition, targetPrice]
      properties:
        symbol:
          type: string
        market:
          type: string
          enum: [PSE, LSE, US, CRYPTO, CUSTOM]
        name:
          type: string
        condition:
          type: string
          enum: [GTE, LTE, EQ, CROSS_UP, CROSS_DOWN]
        targetPrice:
          type: number
        currency:
          type: string
    PriceAlertUpdate:
      type: object
      properties:
        condition:
          type: string
          enum: [GTE, LTE, EQ, CROSS_UP, CROSS_DOWN]
        targetPrice:
          type: number
        status:
          type: string
          enum: [PENDING, FIRED, CLEARED, DISMISSED]

    # ── Notification ───────────────────────────────────────────────────────────
...
    SymbolSearchResult:
      type: object
      required: [symbol, name, market, currency]
      properties:
        symbol:
          type: string
        name:
          type: string
        market:
          type: string
        currency:
          type: string
        exchange:
          type: ["string", "null"]

    # ── Dividend Calendar ──────────────────────────────────────────────────────
...